Add system sequence diagrams for server components
These diagrams describe individual behavioral components of the
database server as well as provide an outline for the inner workings
of the Datastore, Journal, and Database classes.  The functionality
covered is primarily the rotation of the Journal, Datastore, Record
modifications as well as the operations required to start and stop the
server.

These diagrams can be generated using PlantUML.
http://plantuml.sourceforge.net/