Refactor phase-1 document for presentation
This patch addresses the problems with our phase 1 document as
described by the professor and adds automation to the building of the
final pdf. Unused design diagrams were removed from the design folder
for the sake of cleanliness. To compile the latex document and the
images associated with it, use the makefile included in the design
folder.