Add DataController and Journal
This patch adds the required functionality for the DataController and
the scaffolding for the Journal. The DataController has the ability
to process transactions, break those transactions into events, and
then send those events to the journal for recording. It then has the
Datastore act on those events and returns the packaged results.